+++ test/cli/60_htn.sh
@@ -29,35 +29,12 @@ pamela -i "$CODE/test/pamela/biased-coin.pamela" \
        -o "$RESULTS/${NUMBER}_$label" \
        htn
 
-# FUTURE: meaningingfully compare output to rubrique
+if ! diff -u "$dir/${NUMBER}_$label.htn.$format" \
+     "$RESULTS/${NUMBER}_$label.htn.$format"; then
+    exit 1
+fi
 
-# # remove gensym artifacts
-# # NOTE: maps should be sorted by key
-
-# # HTN check ---
-# sed -e 's/\(:[a-z\-]*\)-[0-9][0-9]*/\1-1000/g' \
-#     "$dir/${NUMBER}_$label.htn.$format" > \
-#     "$RESULTS/${NUMBER}_EXPECTED_$label.htn.safe.$format"
-
-# sed -e 's/\(:[a-z\-]*\)-[0-9][0-9]*/\1-1000/g' \
-#     "$RESULTS/${NUMBER}_$label.htn.$format" > \
-#     "$RESULTS/${NUMBER}_$label.htn.safe.$format"
-
-# if ! diff -u "$RESULTS/${NUMBER}_EXPECTED_$label.htn.safe.$format" \
-#      "$RESULTS/${NUMBER}_$label.htn.safe.$format"; then
-#     exit 1
-# fi
-
-# # TPN check ---
-# sed -e 's/\(:[a-z\-]*\)-[0-9][0-9]*/\1-1000/g' \
-#     "$dir/${NUMBER}_$label.tpn.$format" > \
-#     "$RESULTS/${NUMBER}_EXPECTED_$label.tpn.safe.$format"
-
-# sed -e 's/\(:[a-z\-]*\)-[0-9][0-9]*/\1-1000/g' \
-#     "$RESULTS/${NUMBER}_$label.tpn.$format" > \
-#     "$RESULTS/${NUMBER}_$label.tpn.safe.$format"
-
-# if ! diff -u "$RESULTS/${NUMBER}_EXPECTED_$label.tpn.safe.$format" \
-#      "$RESULTS/${NUMBER}_$label.tpn.safe.$format"; then
-#     exit 1
-# fi
+if ! diff -u "$dir/${NUMBER}_$label.tpn.$format" \
+     "$RESULTS/${NUMBER}_$label.tpn.$format"; then
+    exit 1
+fi